Steady progress
Have most of the implementation done with one glaring gotcha, now
working on tests and making sure everything is ship shape per
expectations.

The glaring gotcha is that right now permissions are applied via *the
current global state* of the received transactions. What this means is
that if you add 5 data packets off tx ABC, and someone branches off ABC
and restricts your permissions after you apply your data packets, when
you merge in their permissions change, your packets will no longer be
valid and you'll have to trash them. this is a significant usability
issue and UX issue.

what's next is to build a set of permissions/state based on the previous
transaction chain of the current transaction, not ALL previous
transactions mind you, just the ones you'd find going backwards by
following the `previous_transactions` for each tx. this allows local AND
deterministic permissions for each transaction. this is NOT a simple
change especially keeping performance in mind and will likely require a
lot of clever caching and state management, hence the commit...

so hold on to your butts