Splitting topic state and topics, optimizing state construction
A topic's state can now have multiple copies (via `TopicState`) which is
tracked by stamp_core's new `Dag::apply()` function which manages state
for us so we don't go copying code all over the place. A topic does
still have a final global state, however, which is the result of all
transactions run in order (same mechanism as building an identity).

we now also only reconstruct the state from the DAG if new transactions
branch off of anything other that tail DAG nodes. this means for most
cases we can construct our state in a feed-forward manner without having
to wipe the slate clean and re-run the DAG. it's possible there are more
optimizations here i'm leaving on the table, but for now this is an
acceptable middle ground. perhaps we can squeeze more juice out once we
take a look at snapshots (which is up next).

testing some branching scenarios that make sure this is all working
properly (optimizations and all).